+++ .env
@@ -1,7 +1,7 @@
-GOOGLE_SEARCH_API_KEY = AIzaSyCU6_DO5JWubdbBolU-ciATkPOclNCHTEI
-GOOGLE_SEARCH_CX = 55c1332fff9a840f0
-OPENAI_APIKEY = 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
-MAILGUN_API = f0367b6781e065be1a2cd4c4a28e8b33-db4df449-85ecd538
+GOOGLE_SEARCH_API_KEY=AIzaSyCU6_DO5JWubdbBolU-ciATkPOclNCHTEI
+GOOGLE_SEARCH_CX=55c1332fff9a840f0
+OPENAI_APIKEY=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
+MAILGUN_API=ed4931b3780088a65f343d2995221a67-db4df449-d3a02661
 SDAPI_KEY = sk-eSEHceRyOFkTeuLB2VQuL8OiYM3gwf2wqXEHTUtTmacYJdCj
-MAILGUN_LINK =   https://api.mailgun.net/v3/sandbox0eb00b3b4ea9459faad17ff10f9660df.mailgun.org/messages
-DOMAIN = mailgun@sandbox0eb00b3b4ea9459faad17ff10f9660df.mailgun.org
+MAILGUN_LINK=https://api.mailgun.net/v3/sandbox35481a3344084434a0cd807e4fc5e94b.mailgun.org/messages
+DOMAIN=mailgun@sandbox35481a3344084434a0cd807e4fc5e94b.mailgun.org
