+++ app/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 
 class EmailForm(forms.Form):
-    recipients = forms.CharField(label='Recipient', widget=forms.Textarea)
+    # recipients = forms.CharField(label='Recipient', widget=forms.Textarea)
     topic = forms.CharField(label='Topic')
     imagine = forms.CharField(label='Describe the image you want to see')
     # attachment = forms.FileField(label='Attachment', required=False)
