+++ app/summarize.py
@@ -1,0 +1,66 @@
+import openai
+import requests
+import os
+from googleapiclient.discovery import build
+from datetime import datetime, timedelta
+
+def open_file(filepath):
+    with open(filepath, 'r', encoding='utf-8') as infile:
+        return infile.read()
+        
+def save_file(filepath, content):
+    with open(filepath, 'w', encoding='utf-8') as outfile:
+        outfile.write(content)
+
+
+# Read the Google API key and Custom Search Engine ID from files
+GOOGLE_SEARCH_API_KEY = os.environ.get('GOOGLE_API_KEY')
+GOOGLE_SEARCH_CX = os.environ.get('GOOGLECX_API_ID')
+
+openai.api_key = os.environ.get('OPENAI_API_KEY')
+conversation = []
+
+def fetch_ai_news(topic):
+    search_engine = build("customsearch", "v1", developerKey=GOOGLE_SEARCH_API_KEY)
+    query = f"{topic} site:news.google.com"
+    results = search_engine.cse().list(q=query, cx=GOOGLE_SEARCH_CX, num=6).execute()
+
+    # Extract the titles, snippets, and URLs
+    news_items = [{'title': result['title'], 'snippet': result['snippet'], 'url': result['link']} for result in results['items']]
+    return news_items
+
+
+def summarize_headlines(news_items):
+    summarized_headlines = []
+    for item in news_items:
+        headline = item['title']
+        snippet = item['snippet']
+        url = item['url']
+        summary = chatgpt(f"Please summarize the following headline, snippet and include link(url) for each snippet: {headline} - {snippet} - {url}")
+        summarized_headlines.append(summary)
+    return summarized_headlines
+
+
+
+def save_headlines_to_file(filepath, headlines):
+    with open(filepath, 'w', encoding='utf-8') as outfile:
+        for headline in headlines:
+            outfile.write(headline + "\n")
+
+
+def chatgpt(user_input, temperature=1, frequency_penalty=0.2, presence_penalty=0):
+    global conversation
+
+    conversation.append({"role": "user", "content": user_input})
+    messages_input = conversation.copy()
+
+    completion = openai.ChatCompletion.create(
+        model="gpt-3.5-turbo",
+        temperature=temperature,
+        frequency_penalty=frequency_penalty,
+        presence_penalty=presence_penalty,
+        messages=messages_input)
+
+    chat_response = completion['choices'][0]['message']['content']
+    conversation.append({"role": "assistant", "content": chat_response})
+    return chat_response
