+++ app/templates/temp.html
@@ -20,10 +20,10 @@
             <form method="post" enctype="multipart/form-data" id="emailForm">
                 {% csrf_token %}
                 <!-- form fields here -->
-                <div class="mb-6">
+                <!-- <div class="mb-6">
                     {{ form.recipients.label_tag }} 
                     <input type="text" name="{{ form.recipients.name }}" class="w-full px-3 py-2 border border-gray-900 rounded text-black">
-                </div>
+                </div> -->
                 <div class="mb-6">
                     {{ form.topic.label_tag }} 
                     <input type="text" name="{{ form.topic.name }}" class="w-full px-3 py-2 border border-gray-900 rounded text-black">
