+++ app/views.py
@@ -1,42 +1,65 @@
 from django.shortcuts import render
-from .utils import *
+import os
+from django.conf import settings
+from .forms import EmailForm
+from .mail import *
+from .mail import chatgpt as chat
+from .summarize import *
+import datetime
+import base64
+
+conversation1 = []
+
+api_key = os.environ.get('OPENAI_API_KEY')
+sd_api_key = os.environ.get('SD_API_KEY')
+mailgun_api_key = os.environ.get('MAILGUN_API_KEY')
+
+
+prompt = os.path.join(settings.PROMPT)
+chatbot = os.path.join(settings.CHATBOT)
+
 # Create your views here.
 def fnIndex(request):
     return render(request=request, template_name="home.html")
 
 def fnSendEmail(request):
     if request.method == "POST":
-            recipentEmail = request.POST['recipentEmail']
-            topic = request.POST['topic']
-            emimage_prompt = request.POST['emimage_prompt']
+        form = EmailForm(request.POST, request.FILES)
+        if form.is_valid():
+            # recipients = form.cleaned_data['recipients']
+            topic = form.cleaned_data['topic']
+            imagine = form.cleaned_data['imagine']
+            # attachment = request.FILES.get('attachment')
 
+            # Perform chatgpt and generate_image operations as needed
+            # ...
             headlines = fetch_ai_news(topic)
             summarized_headlines = summarize_headlines(headlines)
-            # # Save the summarized headlines to a text file
-            ainews = save_headlines_to_file(summarized_headlines)
-            print(ainews)
-            with open("prompt1.txt", "r") as fp:
-                prompt1 = fp.read()
-            prompt1 = prompt1.replace("<<AINEWS>>", ainews).replace("\n", "\n\n")
-            conversation1 = []
-            email_content = chatgpt_auto(conversation1, prompt1, prompt1)
-            # Add HTML line breaks
-            email_content = email_content.replace("\n", "<br>")
-            # Get today's date
-            date =datetime.today().strftime("%B %d, %Y")
+            save_headlines_to_file("news_summaries.txt", summarized_headlines)
+            news_file = open_file("news_summaries.txt")
+            print(news_file)
+            prompt1 = open_file(prompt).replace("<<Agent PathFinder>>", news_file).replace("\n", "\n\n")
+            shorten = prompt1[:2000]
+            # Send the email with the generated image attached all emails
+            """For the news"""
+            news_content = chatgpt(api_key, conversation1, summarized_headlines, summarized_headlines)
+            news_content = news_content.replace("\n", "<br>")
+            image_prompt = chat(api_key, conversation1, shorten, imagine)
+            # get today's date
+            today = datetime.date.today().strftime("%B %d, %Y")
+
             # Add the date to the email subject
-            email_subject = "Today's AI News 🤖🔥 " + date
-            # Get the email body
-            email_body = email_content
-            image_prompt_response = chatgpt_auto(conversation1, prompt1, emimage_prompt)
-            if emimage_prompt != "":
+            news_subject = f"Today's {topic} News " + today
+            # get the email body
+            news_body = news_content
+            if imagine != "":
                 try:
-                    image_path = generate_image(image_prompt_response)
-                    sent_result = send_email(recipentEmail, email_subject, email_body, image_path)
+                    image_path = generate_image(sd_api_key, image_prompt)
+                    sent_result = send_email(mailgun_api_key, news_subject, news_body, image_path)
                 except:
                     return render(request=request, template_name="response.html", context={"response": "Failed generate image"})
             else:
-                sent_result = send_email(recipentEmail, email_subject, email_body)
+                sent_result = send_email(mailgun_api_key, news_subject, news_body, image_path)
             if sent_result == True:
                 return render(request=request, template_name="response.html", context={"response": "Your email has been sent successfully."})
             else:
